fix(dialog): label co-varies with impact list; move CascadeImpact to types
